Factor out bitcoin.Config
Since bitcoin has its own specific configuration, it makes sense to give it its own package. We also take the time to pass a more specific path of the config (Extensions.TBTC.BTCRefunds vs Extensions.TBTC) into the function chain.
