+++ config/config.go
@@ -8,7 +8,7 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/keep-network/keep-common/pkg/chain/ethereum"
 	"github.com/keep-network/keep-core/pkg/net/libp2p"
-	eth "github.com/keep-network/keep-ecdsa/pkg/chain"
+	"github.com/keep-network/keep-ecdsa/pkg/chain/bitcoin"
 	"github.com/keep-network/keep-ecdsa/pkg/client"
 	"github.com/keep-network/keep-ecdsa/pkg/ecdsa/tss"
 )
@@ -29,7 +29,19 @@ type Config struct {
 	TSS                    tss.Config
 	Metrics                Metrics
 	Diagnostics            Diagnostics
-	Extensions             eth.Config
+	Extensions             Extensions
+}
+
+// Extensions stores app-specific extensions configuration.
+type Extensions struct {
+	TBTC TBTC
+}
+
+// TBTC stores configuration of application extension responsible for
+// executing signer actions specific for TBTC application.
+type TBTC struct {
+	TBTCSystem string
+	BTCRefunds bitcoin.Config
 }
 
 // SanctionedApplications contains addresses of applications approved by the
