+++ pkg/chain/bitcoin/config.go
@@ -1,0 +1,7 @@
+package bitcoin
+
+// Config stores configuration related to recovering BTC from a closed keep.
+type Config struct {
+	BeneficiaryAddress string
+	MaxFeePerVByte     int32
+}
